rename: 전략타입 범용적으로 사용하기 위해 파일명 수정 및 경로수정 (#156)
