+++ app/(dashboard)/_ui/strategies-item/index.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link'
 
 import classNames from 'classnames/bind'
 
-import { StrategiesModel } from '@/shared/types/strategy-details-data'
+import { StrategiesModel } from '@/shared/types/strategy-data'
 
 import AreaChart from './area-chart'
 import StrategiesSummary from './strategies-summary'
